Done : Article Service class sucessfully created without bugs
Remain :
To convert the DataType -> ResponseEntity<> Type

+++ Medium-backend/src/main/java/com/kanha/Medium_backend/Service/ArticleService.java
@@ -3,13 +3,16 @@ package com.kanha.Medium_backend.Service;
 import com.kanha.Medium_backend.Repository.ArticleRepo;
 import com.kanha.Medium_backend.Repository.UserRepo;
 import com.kanha.Medium_backend.model.Article;
+import com.kanha.Medium_backend.model.Role;
 import com.kanha.Medium_backend.model.User;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
+import org.springframework.security.access.prepost.PreAuthorize;
 import org.springframework.stereotype.Service;
 
 import java.time.LocalDateTime;
+import java.util.ArrayList;
 import java.util.List;
 import java.util.UUID;
 
@@ -50,16 +53,23 @@ public class ArticleService {
     //update the article
 
     public void updateArticle(Article article, UUID id) {
-        Article article1 = articleRepo.getReferenceById(id);
+        Article article1 = articleRepo.findById(id)
+                .orElseThrow(
+                        () ->
+                                new RuntimeException("Article not Existing")
+                );
 
-        if(article1 != null) {
-            article1.setUser(article1.getUser());
+            article1.setUser(article.getUser());
             article1.setTitle(article.getTitle());
             article1.setContent(article.getContent());
             article1.setTags(article.getTags());
-        }
+
         try {
-            articleRepo.save(article);
+            UUID currentUserId = authService.providingUserId();
+            User user = userRepo.findById(currentUserId).get();
+            article1.setUser(user);
+
+            articleRepo.save(article1);
             System.out.println("Successfully updated");
 //            return new ResponseEntity<>("Successfully Updated the Article", HttpStatus.OK);
         } catch (Exception e) {
@@ -69,10 +79,23 @@ public class ArticleService {
     }
 
     public void deleteArticle(UUID id) {
-        articleRepo.deleteById(id);
+        User user = userRepo.findById(authService.providingUserId()).get();
+
+        UUID userStoredInCurrentArticleId = articleRepo.findById(id).get().getUser().getId();
+        if((user.getId().equals(userStoredInCurrentArticleId)) || (user.getRole() == Role.ADMIN)){
+            articleRepo.deleteById(id);
+        }else{
+            System.out.println("Unauthorized");
+        }
     }
 
-    public Article getArticlesById(UUID id) {
-        return articleRepo.getReferenceById(id);
+    //get all the article of that current user by It's id
+    public List<Article> getArticlesofCurrentUser() {
+        //remember that we'll fetch only those articles which is to be returned by 1 own user
+
+        User currLoggedInUser = userRepo.findById(authService.providingUserId()).get();
+
+        //now we'll check the articles written by him
+        return new ArrayList<Article>(currLoggedInUser.getArticles());
     }
 }
